feat: add greying of messages when mentions list open
 - Move mentions list to ConversationFragment
